refactor: move tests folder in root folder
instead of src folder
